Update index.html
So far, we've used tags that stand on their own. However, we sometimes need to specify an attribute, which is information that the tag needs to do its work. For example, to display an image, we need an <img> tag, but we also need a source attribute so the browser knows the location of the image.